+++ examples/basic/server.ts
@@ -1,5 +1,7 @@
 import { createServer } from "@calumet/suamox-hono-adapter";
 
+const port = Number(process.env.PORT) || 3000;
+
 await createServer({
-  port: 3000,
+  port,
 });