+++ examples/basic/src/vite-env.d.ts
@@ -7,5 +7,9 @@ declare module "virtual:pages" {
 
 declare module "virtual:pages/server" {
   export const routes: import("@calumet/suamox").RouteRecord[];
+  export const renderPage: typeof import("@calumet/suamox").renderPage;
+  export const matchRoute: typeof import("@calumet/suamox").matchRoute;
+  export const resolveRouteModule: typeof import("@calumet/suamox").resolveRouteModule;
+  export const RedirectResponse: typeof import("@calumet/suamox").RedirectResponse;
   export default routes;
 }